requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
psycopg2-binary
django-phonenumber-field[phonenumberslite]
django-simple-history
django-guardian
pytz
pillow
